Removed Navigation Drawer and replaced it with Bottom Navigation. Added icons for Bottom Navigation view items.
Added layouts for the user to manage On Call Times by setting specific times and days for the call manager to be active. These layout are not yet linked to the DB or Call Manager Enabling functionality.

Added activity and fragment transitions for smoother UI transitions

Updated ContactDatabase.java to provide a migration to add the new table for on call items